+++ src/helpers/getCaptionProps.ts
@@ -1,5 +1,8 @@
 import { DayPickerProps, CaptionHtmlProps } from '../types';
 
+/**
+ * Return the props for the Caption component.
+ */
 export function getCaptionProps(props: DayPickerProps): CaptionHtmlProps {
   const { styles, classNames } = props;
   return {
